Move dataset types into _types/dataset.py
Also, provide descriptions to the URL fields to disambiguate Dataset and
UnifiedDataset.
