+++ tamr_client/__init__.py
@@ -16,7 +16,14 @@ logging.getLogger(__name__).addHandler(logging.NullHandler())
 # types
 #######
 
-from tamr_client._types import Attribute, AttributeType, SubAttribute, URL
+from tamr_client._types import (
+    Attribute,
+    AttributeType,
+    Dataset,
+    SubAttribute,
+    UnifiedDataset,
+    URL,
+)
 
 # functions
 ###########
@@ -36,7 +43,6 @@ from tamr_client.session import Session
 from tamr_client import session
 
 # datasets
-from tamr_client.dataset import AnyDataset, Dataset
 from tamr_client import dataset
 
 # records
