+++ tamr_client/_types/__init__.py
@@ -15,5 +15,6 @@ from tamr_client._types.attribute import (
     STRING,
     SubAttribute,
 )
+from tamr_client._types.dataset import AnyDataset, Dataset, UnifiedDataset
 from tamr_client._types.json import JsonDict
 from tamr_client._types.url import URL
