+++ tamr_client/dataset/__init__.py
@@ -1,4 +1,2 @@
 from tamr_client.dataset import dataframe, record, unified
-from tamr_client.dataset.dataset import AnyDataset, Dataset
-from tamr_client.dataset.dataset import from_resource_id
-from tamr_client.dataset.dataset import NotFound
+from tamr_client.dataset._dataset import from_resource_id, NotFound
